+++ libs/adapter/file-store/mongo-db.file-store.spec.ts
@@ -1,5 +1,5 @@
+import { FileEntity } from '@pl-oss/domain';
 import { Collection } from 'mongodb';
-import { FileEntity } from './file.entity';
 import { MongoDBFileStore } from './mongo-db.file-store';
 
 jest.mock('mongodb', () => ({